Extract content/metadata logic into concern
This moves the logic to extract content and metadata from
`PublishingApiAction::Publish` to a set of concerns (in anticipation of
moving the logic into `PublishingApiDocument`).
